Add initial work on the ndau chain
Some of the work which still remains to be accomplished:

- testing of any kind
- transactions other than GTValidatorChange
- actually modifying the internal state on GTValidatorChange transactions
- launching the chain in any configuration

Still, go build ./pkg/ndau succeeds without error. It builds,
and that's a good start for a day's work.

One may note that much of this code is a somewhat refactored version
of stuff from the chaos chain. That's true, and it's the benefit of
having built the chaos chain already.