Add tests of the GTVC transaction
In the process of writing these tests, it became apparent that
certain aspects of the code under test needed to change; the
biggest such issue was that the initialize() function needed
to move into the InitChain function. That's a mixed blessing:
on the one hand, things are unified in a way which makes more
sense now, and we don't worry about having two separate init
steps. On the other hand, there's no error type available in
the InitChain function; the only option if something goes wrong
is to panic. Despite Erlang's influences, that still feels like
bad behavior.

Good news is that GTVC transactions are in and tested.