Rework the app.System call to unpack its value into a provided msgp type
This is good from the standpoint of usability, but it required a
non-trivial rewrite of the entire mock system in order to work.

It would be great if delve / vs code displayed more specific type
information; most of the work time on this ended up being spent
tracking down the disconnect where a particular value was serialized
as a wkt.String and the code tried to retrieve it as wkt.Bytes.