Implement concurrency-safe cache of all system variables
This means that instead of reaching out to the chaos chain on an
ad-hoc basis, possibly using different values for different transactions,
we have a stable value which is updated once per block.

In addition to the determinism gains, this also means that the total
volume of requests to the chaos chain is (possibly substantially)
reduced, much improving performance.