Update init scripts and dockerfile
This is just keeping in sync with the improvements implemented for
the chaos node. We now actually check the github fingerprint to
reduce the risk of MITM attacks.

We also now mock up necessary configuration files, or copy the config
if the environment variable $NDAUNODE_CONFIG is set to the path
of an existing config file.