Refactor: pull out config into its own package
This allows us to import the config package from the ndaunode repo
without creating an import cycle. This in turn means we can write
to the config file from `ndau -make-mocks`, which means we can
persist the BPC keys there, which means we can set this thing up
to write PrintMoney transactions.