Move tool config to this package to break import cycle.
Ndaunode needs to have the ndautool config package available so it
can write to the ndautool config file when it makes mocks. Ndautool
depends on ndaunode for a bunch of stuff. Previously, this was an
import cycle between packages which made it a huge pain to upgrade.

By moving the tool config file into this repo, we can ensure that
all the import arrows point in the same direction, which makes things
better.