Fix bug in which accounts vanished from the config file
Got a bit too clever with the accounts map we crate from toml:
on load, we transform a list into a map with two different keys
for each (pointer) value. This means updating either value updates
both, and allows access by means of either. However, we were
initializing that map wrong: the accounts are
actually stored in TOML as a list, and we looped through that list,
storing the pointer to the value.

The problem is that the list variable was simply overwritten every
iteration, not freshly allocated, so we ended up with every key
pointing to a single value, the last in the list.

The solution was to simply create a copy of the variable. The copy
is freshly allocated each time, meaning that the pointers remain
distinct.