WIP: Update to track the metanode which creates only nonempty blocks
This is waiting on two items:

- we need to merge and give version numbers to the various dependencies
  which are currently based on branches, and then update to track the
  appropriate versions in glide.lock
- There's a funny heisenbug going on right now, described below, which
  needs fixing.

Heisenbug: if you run

```sh
go test ./pkg/ndau --run TestAppCanGetFutureValueOnceHeightIsAppropriate
```

it fails, for reasons which are not yet clear.

If you attempt to track it down in a debugger, running

```sh
 dlv test ./pkg/ndau/ -- --test.run TestAppCanGetFutureValueOnceHeightIsAppropriate
```

and set a breakpoint at `system_vars_test.go:73`, you'll see that
`expect` and `value` are in fact identical. It's an interesting one.