Add test of notify tx
This doesn't cover every line of tx_notify.go: there are a few
simple cases where some aspect of accountData is or is not nil
which invalidate the transaction, and it seemed more trouble than
it was worth to write tests for each of those. The big stuff is in.