Update command-line tool to handle transfer scripts properly
There were a few implied tasks: testing this involved writing
a transfer script, and then we needed the ability to select the
keys sent on a per-transaction basis.

A useful validation script for testing is `oAAasUiI`, which decompiles
to `one and not`. This is valid if the first transfer key is set,
regardless of the number of keys sent.