Add test that transaction fees operate properly in RFE txs
The original intent in adding the test to this tx was because
we're going to be removing the TxFeeAccount field from that
transactable soon, so having a test there will help catch any
behavioral errors which creep in.

In implementing this test, though, two distinct behavioral errors
were discovered. Both were fixed for this test, but one was subtle
enough that I think we need to create an equivalent test for every
transaction.