Eliminate actual dependency on oneiro-ndev/signature
We have to keep it as an alias for oneiro-ndev/ndaumath for now,
though, due to a circular dependency. Glide isn't smart enough to
see that we do not and cannot import any code from
oneiro-ndev/chaincode/cmd/chasm into this package, and so adds
oneiro-ndev/signature as a transitive dependency. We can't just
update chaincode first because it depends on this package.