WIP: Add gomobile-compatible interface for transactions
The goal is to write a generator which, when run, inspects the actual
transaction structs, and generates an appropriate file for each.

However, the current work consists of only a proof-of-concept: the
basic shape which the API will take, implemented by hand for the
transfer transaction. This was written to help figure out what
the output for the generator should look like.