Simplify the caching system
This is a simpler and more idiomatic system for filling the cache by using goroutines to dole out the work.

It removes the separate error channel  as it was blocking use of select to no particular benefit. It also eliminates getkv since putting its logic inline actually simplified it.

Note that the result loop now knows how many items are in its stream.