Add ability to generate transaction JSON examples
The mechanism here is kind of weird: running "go generate" on the
pkg builds and runs a little "cmd" package, which writes another
"cmd" package, which is then run, which writes out all the examples.
The reason it's like this is somewhat involved, but it comes down to
the fact that when working with the AST, it's easier to construct
a string which would represent a type as a literal in go source than
to construct the actual corresponding literal.

It all works, though.