move codegen to generator project
When generation lives in this project, it is very hard to iterate
when incomplete or incorrect generation leaves this project in a
state which doesn't compile. It was therefore easiest to move that
to a different project.

We'll leave generated code in the repo to be checked in as normal.