Prevalidate and GetOrderChainInfo get SIB from the official source
This involved a substantial simplification of GetOrderChainInfo, as
we got rid of a bunch of defaults and replaced them with real values,
but we simply don't have real values for each default present. Eliminated
the default values on the theory it's better to have less information
and more confidence in it than the reverse.

Also added a tool.GetSIB method.