Fix some broken tests
Turns out that we'd previously used NodeRewardWinner as if it were
optional, setting a flag to indicate if it were present. However,
we didn't actually represent that in the typesystem. This caused
the generator to omit the flag in the generated code, which in turn
broke the tests, because the generated code didn't know to ignore
it if it were not present. Happily, a bunch of tests did catch this.

Fix: change the type to a pointer so the nil case is handled, and
adjust the code affected by the type change.