Add test which fails for this bug
It doesn't help diagnose why the app's block time is sometimes
not updated by new blocks arriving, but it does at least demonstrate
the smaller of the bugs in that issue: locks aren't being cleared
when they should be. More importantly, it gives us a straightforward
way to get this case in a debugger.