propagate tx value struct instead of raw values
It took several hours to figure out why the tests weren't passing.
There were two culprits. The first was a relatively simple error:
an important struct's fields were private, meaning that they weren't
serialized to MSGP, which meant that deserialization zeroed them, which
meant that it wasn't persisted properly. The other was harder:
it turned out that somehow, at some point, a redis instance failed
to die, meaning that it wasn't empty, which meant that the test's
preconditions were wrong.

That's all done, and _finally_ I can move forward with updating the API.