Refactor Cfg: retain a persistent node connection
This should provide a minor performance improvement, and definitely
saves a lot of redundant code, but the main reason for the change
is so that we can construct a config object which mocks out TM entirely:
just implement `TMClient` with your mock type and you're golden.
This should make it much easier to write good tests.

This refactor removes two test files. The only tests they contained
were wholly dependent on the behavior of partially-initialized objects.
One should never attempt to use a partially-initialized object anyway,
so it seemed better to not test the behavior thereof.