cause all tests to pass
This involved nontrivial changes, but the end result is that we now
have a comprehensive test of the API: changing any tested endpoints
and the data structures returned therefrom will break these tests.

This also removes the nodes from the SDK, because the existing
ndauapi implementation is broken. If nobody's noticed so far,
chances seem good that nobody is attempting to use those endpoints
anyway.