fix the deferred closure heap overwrite problem
The motivation for https://github.com/oneiro-ndev/ndau/commit/ab3f086d0a4a9ed3c465ae125620b60fa99e919d
was that a bug was discovered as I was going on vacation, and the
person discovering the bug didn't know how to fix it.

This commit adds the appropriate fix, and a note explaining how
it works.