express total stake and total delegation as ratios in goodness func
I'm fairly well convinced that this is a bad idea that we shouldn't do,
but the feature has been requested, so here it is.

The problem is that without this change, it is possible to directly
compare stake and delegation: they both have units of
ndau. With this change, both those numbers are now unitless ratios;
we have lost information.

Computationally, I've added a cache, but we still need to loop over
the entire address space at least once per block, because node goodnesses
are calculated every block. This will slow things down.

Mathematically, it's not a no-op, but it's similar: for a given state,
the goodness ratios between each node pair are identical with and without
this change.